requirements.txt
pandas==2.2.2
numpy==1.24.3
matplotlib==3.7.1
modal==0.62.109
Pillow==10.3.0
python-dotenv==1.0.0
opencv-python-headless==4.10.0.84
transformers==4.40.0
fastapi==0.111.0
tensorflow~=2.9.1
